Move block placement default to HarvestTool
