+++ src/main/java/tconstruct/library/tools/HarvestTool.java
@@ -5,6 +5,8 @@ import tconstruct.library.TConstructRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
@@ -137,4 +139,82 @@ public abstract class HarvestTool extends ToolCore
     protected abstract Material[] getEffectiveMaterials ();
 
     protected abstract String getHarvestType ();
+    
+    //Right-click
+    @Override
+    public boolean onItemUse (ItemStack stack, EntityPlayer player, World world, int x, int y, int z, int side, float clickX, float clickY, float clickZ)
+    {
+        /*if (world.isRemote)
+            return true;*/
+
+        boolean used = false;
+        int hotbarSlot = player.inventory.currentItem;
+        int itemSlot = hotbarSlot == 0 ? 8 : hotbarSlot + 1;
+        ItemStack nearbyStack = null;
+
+        if (hotbarSlot < 8)
+        {
+            nearbyStack = player.inventory.getStackInSlot(itemSlot);
+            if (nearbyStack != null)
+            {
+                Item item = nearbyStack.getItem();
+                if (item instanceof ItemBlock)
+                {
+                    int posX = x;
+                    int posY = y;
+                    int posZ = z;
+                    int playerPosX = (int) Math.floor(player.posX);
+                    int playerPosY = (int) Math.floor(player.posY);
+                    int playerPosZ = (int) Math.floor(player.posZ);
+                    if (side == 0)
+                    {
+                        --posY;
+                    }
+
+                    if (side == 1)
+                    {
+                        ++posY;
+                    }
+
+                    if (side == 2)
+                    {
+                        --posZ;
+                    }
+
+                    if (side == 3)
+                    {
+                        ++posZ;
+                    }
+
+                    if (side == 4)
+                    {
+                        --posX;
+                    }
+
+                    if (side == 5)
+                    {
+                        ++posX;
+                    }
+                    if (posX == playerPosX && (posY == playerPosY || posY == playerPosY + 1 || posY == playerPosY - 1) && posZ == playerPosZ)
+                    {
+                        return false;
+                    }
+
+                    used = item.onItemUse(nearbyStack, player, world, x, y, z, side, clickX, clickY, clickZ);
+                    if (nearbyStack.stackSize < 1)
+                    {
+                        nearbyStack = null;
+                        player.inventory.setInventorySlotContents(itemSlot, null);
+                    }
+                }
+            }
+        }
+
+        /*if (used) //Update client
+        {
+            Packet103SetSlot packet = new Packet103SetSlot(player.openContainer.windowId, itemSlot, nearbyStack);
+            ((EntityPlayerMP)player).playerNetServerHandler.sendPacketToPlayer(packet);
+        }*/
+        return used;
+    }
 }
